pyproject.toml
[project]
name = "learn-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.12",
    "langchain-ollama>=0.3.8",
    "python-dotenv>=1.1.1",
]
